+++ .vscode/launch.json
@@ -28,15 +28,8 @@
                 ]
             },
             "classPaths": [
-                "$Auto",
-                "${workspaceFolder}/plugin/hello-world/target/classes",
-                "${workspaceFolder}/plugin/cpg/target/classes",
-                "${workspaceFolder}/plugin/ra/target/classes",
-                "${workspaceFolder}/plugin/sdc/target/classes",
-                "${workspaceFolder}/plugin/utility/target/classes",
-					 "${workspaceFolder}/plugin/devtools/target/classes",
-					 "${workspaceFolder}/plugin/cql/target/classes",
-					 "${workspaceFolder}/plugin/security/target/classes"
+               "$Auto",
+					"$Runtime"
             ]
         },
         {
