+++ plugin/cpg/pom.xml
@@ -1,12 +1,10 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
-			xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-			xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 	<parent>
 		<groupId>org.opencds.cqf.ruler</groupId>
-		<artifactId>cqf-ruler</artifactId>
+		<artifactId>cqf-ruler-plugin</artifactId>
 		<version>0.5.0-SNAPSHOT</version>
-		<relativePath>../..</relativePath>
 	</parent>
 
 	<artifactId>cqf-ruler-plugin-cpg</artifactId>
@@ -21,7 +19,13 @@
 			<groupId>org.opencds.cqf.ruler</groupId>
 			<artifactId>cqf-ruler-server</artifactId>
 			<version>0.5.0-SNAPSHOT</version>
+			<scope>provided</scope>
 			<classifier>classes</classifier>
+		</dependency>
+		<dependency>
+			<groupId>org.opencds.cqf.ruler</groupId>
+			<artifactId>cqf-ruler-plugin-api</artifactId>
+			<version>0.5.0-SNAPSHOT</version>
 			<scope>provided</scope>
 		</dependency>
 		<dependency>
@@ -58,7 +62,7 @@
 		</dependency>
 		<dependency>
 			<groupId>org.opencds.cqf.ruler</groupId>
-			<artifactId>cqf-ruler-plugin-test-utility</artifactId>
+			<artifactId>cqf-ruler-test</artifactId>
 			<version>0.5.0-SNAPSHOT</version>
 			<scope>test</scope>
 		</dependency>
