+++ plugin/cpg/src/test/java/org/opencds/cqf/ruler/plugin/cpg/r4/provider/LibraryEvaluationProviderIT.java
@@ -16,7 +16,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.opencds.cqf.ruler.Application;
 import org.opencds.cqf.ruler.plugin.cpg.CpgConfig;
 import org.opencds.cqf.ruler.plugin.cpg.CpgProperties;
-import org.opencds.cqf.ruler.plugin.testutility.ResolutionUtilities;
+import org.opencds.cqf.ruler.test.ResolutionUtilities;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
