+++ plugin/cr/src/test/java/org/opencds/cqf/ruler/plugin/cr/r4/ExpressionEvaluationIT.java
@@ -24,7 +24,7 @@ import org.opencds.cqf.ruler.plugin.cql.CqlConfig;
 import org.opencds.cqf.ruler.plugin.cr.CrConfig;
 import org.opencds.cqf.ruler.plugin.devtools.DevToolsConfig;
 import org.opencds.cqf.ruler.plugin.devtools.r4.CodeSystemUpdateProvider;
-import org.opencds.cqf.ruler.plugin.testutility.IServerSupport;
+import org.opencds.cqf.ruler.test.IServerSupport;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
