+++ plugin/dev-tools/src/test/java/org/opencds/cqf/ruler/plugin/devtools/dstu3/CacheValueSetsProviderIT.java
@@ -21,9 +21,9 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mockito;
 import org.opencds.cqf.ruler.Application;
 import org.opencds.cqf.ruler.plugin.devtools.DevToolsConfig;
-import org.opencds.cqf.ruler.plugin.testutility.IServerSupport;
 import org.opencds.cqf.ruler.plugin.utility.ClientUtilities;
 import org.opencds.cqf.ruler.plugin.utility.IdUtilities;
+import org.opencds.cqf.ruler.test.IServerSupport;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
