+++ plugin/ra/src/test/java/org/opencds/cqf/ruler/plugin/ra/r4/ReportProviderIT.java
@@ -18,7 +18,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.opencds.cqf.ruler.Application;
 import org.opencds.cqf.ruler.plugin.ra.RAConfig;
 import org.opencds.cqf.ruler.plugin.ra.RAProperties;
-import org.opencds.cqf.ruler.plugin.testutility.ResolutionUtilities;
+import org.opencds.cqf.ruler.test.ResolutionUtilities;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
@@ -37,7 +37,7 @@ import ca.uhn.fhir.rest.server.exceptions.ResourceNotFoundException;
 @ActiveProfiles("test")
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT, classes = { Application.class,
 		RAConfig.class }, properties = { "hapi.fhir.fhir_version=r4" })
-public class ReportProviderIT implements org.opencds.cqf.ruler.plugin.testutility.ClientUtilities, ResolutionUtilities,
+public class ReportProviderIT implements org.opencds.cqf.ruler.test.ClientUtilities, ResolutionUtilities,
 		org.opencds.cqf.ruler.plugin.utility.ClientUtilities {
 
 	private IGenericClient ourClient;
