+++ plugin/sdc/src/test/java/org/opencds/cqf/ruler/plugin/sdc/r4/ExtractProviderIT.java
@@ -15,7 +15,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.opencds.cqf.ruler.Application;
 import org.opencds.cqf.ruler.plugin.sdc.SDCConfig;
 import org.opencds.cqf.ruler.plugin.sdc.SDCProperties;
-import org.opencds.cqf.ruler.plugin.testutility.ResolutionUtilities;
+import org.opencds.cqf.ruler.test.ResolutionUtilities;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
