+++ test/pom.xml
@@ -5,28 +5,14 @@
 		<groupId>org.opencds.cqf.ruler</groupId>
 		<artifactId>cqf-ruler</artifactId>
 		<version>0.5.0-SNAPSHOT</version>
-		<relativePath>../..</relativePath>
 	</parent>
 
-	<artifactId>cqf-ruler-plugin-test-utility</artifactId>
+	<artifactId>cqf-ruler-test</artifactId>
 	<dependencies>
 		<dependency>
 			<groupId>org.opencds.cqf.ruler</groupId>
 			<artifactId>cqf-ruler-external</artifactId>
 			<version>0.5.0-SNAPSHOT</version>
-			<scope>provided</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.opencds.cqf.ruler</groupId>
-			<artifactId>cqf-ruler-server</artifactId>
-			<version>0.5.0-SNAPSHOT</version>
-			<classifier>classes</classifier>
-			<scope>provided</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-autoconfigure</artifactId>
-			<scope>provided</scope>
 		</dependency>
 		<dependency>
 			<groupId>org.junit.jupiter</groupId>
